feat: enhance metrics read functionality with CSV export support
