+++ utils/index.ts
@@ -32,3 +32,4 @@ export * from './consts.js';
 export * from './snake-to-camel.js';
 export * from './timestamp.js';
 export * from './wallet-connect.js';
+export * from './csv-file.js';
